refactor(layout): structure process layout pipeline
Extract process-scope orchestration into a functional pipeline and define its internal context contracts. Keep artifact orchestration and layout state ownership explicit without changing generated geometry.
