+++ lib/Layouter.js
@@ -4,10 +4,9 @@ import { DiFactory } from './di/DiFactory.js';
 import { getDefaultSize, is } from './di/DiUtil.js';
 import { LayoutError } from './LayoutError.js';
 import { LayoutWarning } from './LayoutWarning.js';
-import { HORIZONTAL_GAP, VERTICAL_GAP, SUB_PROCESS_PADDING, PARTICIPANT_HEADER_WIDTH, MIN_PARTICIPANT_WIDTH, MIN_PARTICIPANT_HEIGHT, MIN_SUB_PROCESS_WIDTH, MIN_SUB_PROCESS_HEIGHT, EXPANDED_SUBPROCESS_ANNOTATION_CLEARANCE, EXPANDED_SUBPROCESS_LABEL_HEIGHT, MAX_ARTIFACT_SEARCH_OFFSET } from './layout/Constants.js';
+import { HORIZONTAL_GAP, VERTICAL_GAP } from './layout/Constants.js';
 import {
   isArtifact,
-  isExteriorArtifact,
   isSupportedVisualElement,
   isSupportedVisualConnection,
   getExpandedIds
@@ -18,37 +17,16 @@ import {
   normalizeLayout,
   translateLayout,
   getExtents,
-  getParticipantContentExtents,
   hasParticipantContent,
-  getShapeExtents,
   getExpandedChildEdges,
   getExpandedChildShapes
 } from './layout/LayoutUtil.js';
 import {
-  validateSequenceFlows,
   validateMessageFlows,
-  validateBoundaryEvents,
-  validateLinks,
   validateInputVisuals,
   validateParseWarnings
 } from './layout/Validation.js';
-import { createSemanticPolicy, assignRanks, edgePriority } from './layout/SemanticPolicy.js';
-import {
-  placeRecords,
-  clearBoundaryHandlerExits,
-  compactSemanticBands,
-  packComponents,
-  applyLaneMembership,
-  flattenLanes,
-  placeBoundaryEvents
-} from './layout/ShapePlacement.js';
-import { routeConnection } from './layout/SequenceFlowRouter.js';
-import {
-  findContainingArtifactContainers,
-  artifactSizeCandidates,
-  findArtifactPlacement,
-  routeArtifactAssociation
-} from './layout/ArtifactLayouter.js';
+import { flattenLanes } from './layout/ShapePlacement.js';
 import {
   sizeAndPositionParticipantsFromMessageAnchors,
   includeResizableParticipantMessageDocks,
@@ -59,12 +37,15 @@ import {
   getMessageObstacles,
   routeMessageFlows
 } from './layout/CollaborationLayouter.js';
-import {
-  layoutExternalLabels,
-  needsExpandedSubProcessTitleClearance
-} from './layout/LabelLayouter.js';
+import { layoutExternalLabels } from './layout/LabelLayouter.js';
 import { layoutGroups } from './layout/GroupLayouter.js';
 import { finalizeLayoutConnections } from './layout/ConnectionDocking.js';
+import { placeArtifacts } from './layout/ArtifactLayoutStage.js';
+import {
+  createLayoutRecord,
+  getParticipantContainerBounds,
+  layoutProcessScope
+} from './layout/ProcessLayoutPipeline.js';
 import { emitLayout } from './di/LayoutEmitter.js';
 
 /**
@@ -96,10 +77,16 @@ export class Layouter {
     this.expandedIds = getExpandedIds(definitions, root);
     validateInputVisuals(definitions, root);
     definitions.diagrams = [];
+    let layout;
 
-    const layout = is(root, 'bpmn:Collaboration')
-      ? this.layoutCollaboration(root)
-      : this.layoutScope(root);
+    if (is(root, 'bpmn:Collaboration')) {
+      layout = this.layoutCollaboration(root);
+    } else {
+      const result = layoutProcessScope(root, { expandedIds: this.expandedIds });
+
+      layout = result.layout;
+      this.warnings.push(...result.warnings);
+    }
 
     normalizeLayout(layout);
     finalizeLayoutConnections(layout, true);
@@ -177,15 +164,17 @@ export class Layouter {
         continue;
       }
 
-      const processLayout = this.layoutScope(
-        process,
-        true,
+      const processResult = layoutProcessScope(process, {
+        expandedIds: this.expandedIds,
+        participantProcess: true,
         messageFlowEndpointDirections
-      );
-      const participantRect = this.getParticipantContainerBounds(process, processLayout);
+      });
+      const processLayout = processResult.layout;
+      const participantRect = getParticipantContainerBounds(process, processLayout);
       const dx = -participantRect.x;
       const dy = -participantRect.y;
 
+      this.warnings.push(...processResult.warnings);
       translateLayout(processLayout, dx, dy);
 
       layout.shapes.set(participant, bounds(
@@ -338,513 +327,20 @@ export class Layouter {
     const groups = artifacts.filter(element => is(element, 'bpmn:Group'));
     const artifactRecords = artifacts
       .filter(element => isArtifact(element) && !is(element, 'bpmn:Group'))
-      .map((element, index) => this.createRecord(element, index));
+      .map((element, index) => createLayoutRecord(element, index, this.expandedIds));
     const associations = artifacts.filter(element => is(element, 'bpmn:Association'));
 
-    this.placeArtifacts(
-      artifactRecords,
+    placeArtifacts({
+      records: artifactRecords,
       associations,
       layout,
-      [],
-      new Set(),
-      collaboration.participants.length === 1,
-      collaboration.participants.length !== 1
-    );
+      reservedVerticalEndpointDirections: new Map(),
+      avoidParticipantInterior: collaboration.participants.length === 1,
+      preferParticipantSides: collaboration.participants.length !== 1
+    });
     this.warnings.push(...layoutGroups(groups, layout));
 
     return layout;
-  }
-
-  layoutScope(
-      scope,
-      participantProcess = false,
-      messageFlowEndpointDirections = new Map()) {
-    const layout = createLayout(scope);
-    const flowElements = scope.flowElements || [];
-    const artifacts = scope.artifacts || [];
-    const groups = artifacts.filter(element => is(element, 'bpmn:Group'));
-    const sequenceFlows = flowElements.filter(element => is(element, 'bpmn:SequenceFlow'));
-    const dataAssociations = flowElements.flatMap(element => [
-      ...(element.dataInputAssociations || []),
-      ...(element.dataOutputAssociations || [])
-    ]);
-    const associations = [ ...flowElements, ...artifacts ]
-      .filter(element => is(element, 'bpmn:Association'))
-      .concat(dataAssociations);
-    const nodeElements = [ ...new Set([
-      ...flowElements.filter(element => {
-        return !is(element, 'bpmn:SequenceFlow') &&
-          !is(element, 'bpmn:Association') &&
-          !is(element, 'bpmn:Group') &&
-          !is(element, 'bpmn:DataObject');
-      }),
-      ...artifacts.filter(element => {
-        return isArtifact(element) && !is(element, 'bpmn:Group');
-      })
-    ]) ];
-
-    const records = nodeElements.map((element, index) => this.createRecord(element, index));
-
-    const recordsByElement = new Map(records.map(record => [ record.element, record ]));
-
-    for (const record of records) {
-      if (!is(record.element, 'bpmn:SubProcess')) {
-        continue;
-      }
-
-      record.child = this.layoutScope(
-        record.element,
-        false,
-        messageFlowEndpointDirections
-      );
-
-      if (record.expanded) {
-        const childExtents = getExtents(record.child);
-        record.size = {
-          width: Math.max(MIN_SUB_PROCESS_WIDTH, childExtents.width + 2 * SUB_PROCESS_PADDING),
-          height: Math.max(MIN_SUB_PROCESS_HEIGHT, childExtents.height + 2 * SUB_PROCESS_PADDING)
-        };
-      }
-
-      layout.children.push(record.child);
-    }
-
-    validateSequenceFlows(sequenceFlows, recordsByElement, scope);
-    validateBoundaryEvents(records, recordsByElement, scope);
-    validateLinks(records, scope);
-
-    const graphRecords = records.filter(record => !record.isBoundary && !record.isArtifact &&
-      !(is(record.element, 'bpmn:SubProcess') && record.element.triggeredByEvent));
-    const graphSet = new Set(graphRecords.map(record => record.element));
-    const graphEdges = sequenceFlows.filter(flow => graphSet.has(flow.sourceRef) &&
-      graphSet.has(flow.targetRef));
-    const boundaryEdges = sequenceFlows.filter(flow => is(flow.sourceRef, 'bpmn:BoundaryEvent'));
-
-    const policy = createSemanticPolicy(scope, graphRecords, graphEdges, boundaryEdges, records);
-    const ranks = assignRanks(graphRecords, graphEdges, boundaryEdges, policy);
-    policy.backEdges = ranks.backEdges;
-    compactSemanticBands(graphRecords, graphEdges, boundaryEdges, ranks, policy);
-    placeRecords(graphRecords, ranks, policy);
-    clearBoundaryHandlerExits(graphRecords, boundaryEdges, recordsByElement, policy);
-    packComponents(scope, graphRecords, graphEdges, boundaryEdges);
-    applyLaneMembership(scope, graphRecords, graphEdges, policy, layout);
-    placeBoundaryEvents(records, recordsByElement, layout);
-
-    for (const record of graphRecords) {
-      layout.shapes.set(record.element, record.bounds);
-    }
-
-    for (const record of records.filter(record => record.isBoundary)) {
-      layout.shapes.set(record.element, record.bounds);
-    }
-
-    this.placeExpandedChildren(records, layout);
-    this.routeSequenceFlows(sequenceFlows, layout, policy);
-    this.placeEventSubProcesses(records, layout);
-    if (participantProcess) {
-      const interiorArtifacts = records.filter(record => {
-        return record.isArtifact && !isExteriorArtifact(record.element);
-      });
-      const exteriorArtifacts = records.filter(record => {
-        return record.isArtifact && isExteriorArtifact(record.element);
-      });
-
-      this.placeArtifacts(
-        interiorArtifacts,
-        associations,
-        layout,
-        [],
-        messageFlowEndpointDirections
-      );
-
-      const participantBounds = this.getParticipantContainerBounds(scope, layout);
-
-      this.placeArtifacts(exteriorArtifacts, associations, layout, [ {
-        rect: participantBounds,
-        containsOwner: true,
-        participant: true
-      } ], messageFlowEndpointDirections);
-    } else {
-      this.placeArtifacts(
-        records,
-        associations,
-        layout,
-        [],
-        messageFlowEndpointDirections
-      );
-    }
-    this.warnings.push(...layoutGroups(groups, layout));
-
-    return layout;
-  }
-
-  createRecord(element, index) {
-    const size = getDefaultSize(element);
-
-    if (!size || !isSupportedVisualElement(element)) {
-      throw new LayoutError(
-        'UNSUPPORTED_ELEMENT',
-        element.id,
-        `Cannot generate DI for visual BPMN element "${element.$type}".`
-      );
-    }
-
-    return {
-      element,
-      index,
-      size,
-      isBoundary: is(element, 'bpmn:BoundaryEvent'),
-      isArtifact: isArtifact(element),
-      expanded: is(element, 'bpmn:SubProcess') && this.expandedIds.has(element.id),
-      child: null
-    };
-  }
-
-  getParticipantContainerBounds(process, layout) {
-    const extents = getParticipantContentExtents(layout);
-    const hasLanes = flattenLanes(process.laneSets || []).length > 0;
-    const leadingPadding = hasLanes
-      ? PARTICIPANT_HEADER_WIDTH
-      : PARTICIPANT_HEADER_WIDTH + SUB_PROCESS_PADDING;
-    const trailingPadding = hasLanes ? 0 : SUB_PROCESS_PADDING;
-    const verticalPadding = hasLanes ? 0 : SUB_PROCESS_PADDING;
-    const width = Math.max(
-      MIN_PARTICIPANT_WIDTH,
-      extents.width + leadingPadding + trailingPadding
-    );
-    const height = Math.max(
-      MIN_PARTICIPANT_HEIGHT,
-      extents.height + 2 * verticalPadding
-    );
-
-    return bounds(
-      extents.minX - leadingPadding,
-      extents.minY - verticalPadding,
-      width,
-      height
-    );
-  }
-
-  placeExpandedChildren(records) {
-    for (const record of records) {
-      if (!record.expanded || !record.child || !record.bounds) {
-        continue;
-      }
-      const extents = getExtents(record.child);
-      translateLayout(
-        record.child,
-        record.bounds.x + SUB_PROCESS_PADDING - extents.minX,
-        record.bounds.y + SUB_PROCESS_PADDING - extents.minY
-      );
-      record.child.emitInParent = true;
-
-      if (needsExpandedSubProcessTitleClearance(
-        record.element,
-        record.bounds,
-        record.child
-      )) {
-        translateLayout(record.child, 0, EXPANDED_SUBPROCESS_LABEL_HEIGHT);
-      }
-    }
-  }
-
-  placeEventSubProcesses(records, layout) {
-    const eventSubProcesses = records.filter(record => {
-      return is(record.element, 'bpmn:SubProcess') && record.element.triggeredByEvent;
-    });
-
-    if (!eventSubProcesses.length) {
-      return;
-    }
-
-    let nextEventSubProcessY = getExtents(layout).maxY + VERTICAL_GAP;
-
-    for (const record of eventSubProcesses) {
-      if (!record.expanded) {
-        record.bounds = bounds(0, nextEventSubProcessY, record.size.width, record.size.height);
-        layout.shapes.set(record.element, record.bounds);
-        nextEventSubProcessY += record.size.height + VERTICAL_GAP;
-        continue;
-      }
-
-      const extents = getExtents(record.child);
-      const width = Math.max(MIN_SUB_PROCESS_WIDTH, extents.width + 2 * SUB_PROCESS_PADDING);
-      const height = Math.max(MIN_SUB_PROCESS_HEIGHT, extents.height + 2 * SUB_PROCESS_PADDING);
-
-      record.bounds = bounds(0, nextEventSubProcessY, width, height);
-      layout.shapes.set(record.element, record.bounds);
-      translateLayout(
-        record.child,
-        record.bounds.x + SUB_PROCESS_PADDING - extents.minX,
-        record.bounds.y + SUB_PROCESS_PADDING - extents.minY
-      );
-      record.child.emitInParent = true;
-      nextEventSubProcessY += height + VERTICAL_GAP;
-    }
-
-  }
-
-  placeArtifacts(
-      records,
-      associations,
-      layout,
-      additionalBoundaryContainers = [],
-      reservedVerticalEndpointDirections = new Map(),
-      avoidParticipantInterior = false,
-      preferParticipantSides = true) {
-    const artifactRecords = records.filter(record => record.isArtifact);
-    const graphShapes = new Map([
-      ...layout.shapes,
-      ...getExpandedChildShapes(layout)
-    ]);
-    const graphElements = new Set(graphShapes.keys());
-    const owners = new Map();
-    const graphObstacles = [ ...graphShapes.entries() ].filter(([ element ]) => {
-      return !is(element, 'bpmn:Lane') &&
-        !is(element, 'bpmn:Participant') &&
-        !isArtifact(element);
-    }).map(([ element, rect ]) => ({ element, rect }));
-    const graphExtents = getShapeExtents(graphObstacles);
-    const placementExtents = getShapeExtents([ ...graphShapes.entries() ]
-      .filter(([ element ]) => !isArtifact(element))
-      .map(([ element, rect ]) => ({ element, rect })));
-    const currentArtifacts = new Set(artifactRecords.map(record => record.element));
-    const annotatedMessageEndpoints = new Map();
-
-    for (const association of associations) {
-      const endpoints = [
-        ...(Array.isArray(association.sourceRef)
-          ? association.sourceRef
-          : [ association.sourceRef ]),
-        association.targetRef
-      ];
-      const annotation = endpoints.find(endpoint => {
-        return currentArtifacts.has(endpoint) &&
-          is(endpoint, 'bpmn:TextAnnotation');
-      });
-      const owner = endpoints.find(endpoint => endpoint !== annotation);
-
-      if (annotation && reservedVerticalEndpointDirections.has(owner)) {
-        annotatedMessageEndpoints.set(
-          owner,
-          reservedVerticalEndpointDirections.get(owner)
-        );
-      }
-    }
-
-    const graphRoutes = collectArtifactObstacleRoutes(
-      layout,
-      annotatedMessageEndpoints,
-      graphShapes
-    );
-    const placedArtifacts = [ ...graphShapes.entries() ]
-      .filter(([ element ]) => {
-        return isArtifact(element) && !currentArtifacts.has(element);
-      })
-      .map(([ element, rect ]) => ({
-        element,
-        rect,
-        annotationClearance: 0
-      }));
-
-    for (const association of associations) {
-      const endpoints = [
-        ...(Array.isArray(association.sourceRef) ? association.sourceRef : [ association.sourceRef ]),
-        association.targetRef
-      ];
-      const artifact = endpoints.find(endpoint => artifactRecords.some(record => record.element === endpoint));
-      const owner = is(association, 'bpmn:DataAssociation')
-        ? association.$parent
-        : endpoints.find(endpoint => endpoint !== artifact && graphElements.has(endpoint));
-
-      if (!artifact || !owner || !graphElements.has(owner)) {
-        continue;
-      }
-
-      if (!owners.has(artifact)) {
-        owners.set(artifact, []);
-      }
-
-      owners.get(artifact).push({
-        association,
-        owner,
-        ownerBounds: graphShapes.get(owner)
-      });
-    }
-
-    for (const references of owners.values()) {
-      const counts = new Map();
-      const indices = new Map();
-
-      for (const reference of references) {
-        counts.set(reference.owner, (counts.get(reference.owner) || 0) + 1);
-      }
-
-      for (const reference of references) {
-        const index = indices.get(reference.owner) || 0;
-
-        reference.ownerConnectionIndex = index;
-        reference.ownerConnectionCount = counts.get(reference.owner);
-        indices.set(reference.owner, index + 1);
-      }
-    }
-
-    artifactRecords.sort((a, b) => {
-      const aReferences = owners.get(a.element)?.length || 0;
-      const bReferences = owners.get(b.element)?.length || 0;
-      const aArea = artifactSizeCandidates(a.element)[0];
-      const bArea = artifactSizeCandidates(b.element)[0];
-
-      return bReferences - aReferences ||
-        bArea.width * bArea.height - aArea.width * aArea.height ||
-        a.index - b.index;
-    });
-
-    for (const record of artifactRecords) {
-      const references = owners.get(record.element) || [];
-      const ownerBounds = references.length
-        ? graphShapes.get(references[0].owner)
-        : null;
-      const owner = references[0]?.owner;
-      const enclosingSubProcesses = ownerBounds
-        ? [ ...graphShapes.entries() ].filter(([ element, rect ]) => {
-          const centerX = ownerBounds.x + ownerBounds.width / 2;
-          const centerY = ownerBounds.y + ownerBounds.height / 2;
-
-          return element !== owner &&
-            is(element, 'bpmn:SubProcess') &&
-            centerX >= rect.x &&
-            centerX <= rect.x + rect.width &&
-            centerY >= rect.y &&
-            centerY <= rect.y + rect.height;
-        })
-        : [];
-      const subProcessContainer = enclosingSubProcesses
-        .map(([ , rect ]) => rect)
-        .sort((a, b) => a.width * a.height - b.width * b.height)[0];
-      const containingContainers = ownerBounds
-        ? findContainingArtifactContainers(ownerBounds, graphShapes)
-        : [];
-      const container = containingContainers[0];
-      const boundaryContainers = [
-        ...additionalBoundaryContainers,
-        ...[ ...graphShapes.entries() ]
-          .filter(([ element ]) => {
-            return is(element, 'bpmn:Lane') ||
-              is(element, 'bpmn:Participant') ||
-              is(element, 'bpmn:SubProcess');
-          })
-          .map(([ element, rect ]) => ({
-            rect,
-            containsOwner: containingContainers.includes(rect),
-            participant: is(element, 'bpmn:Participant')
-          }))
-      ];
-      const processContainer = subProcessContainer || (
-        !container && isExteriorArtifact(record.element)
-          ? bounds(
-            graphExtents.minX - SUB_PROCESS_PADDING,
-            graphExtents.minY - SUB_PROCESS_PADDING,
-            graphExtents.width + 2 * SUB_PROCESS_PADDING,
-            graphExtents.height + 2 * SUB_PROCESS_PADDING
-          )
-          : null
-      );
-      const annotationClearance = subProcessContainer &&
-        is(record.element, 'bpmn:TextAnnotation')
-        ? EXPANDED_SUBPROCESS_ANNOTATION_CLEARANCE
-        : 0;
-      const obstacles = graphObstacles.filter(({ element }) => {
-        return !enclosingSubProcesses.some(([ subProcess ]) => subProcess === element);
-      });
-      record.associationObstacles = obstacles;
-      const placement = findArtifactPlacement(
-        record.element,
-        {
-          ownerBounds,
-          references,
-          sizes: artifactSizeCandidates(record.element),
-          obstacles,
-          routes: graphRoutes,
-          occupied: placedArtifacts,
-          container,
-          processContainer,
-          extents: placementExtents,
-          annotationClearance,
-          boundaryContainers,
-          avoidParticipantInterior,
-          preferParticipantSides,
-          participantInteriorPreference: (
-            isExteriorArtifact(record.element) &&
-            references.some(({ owner }) => {
-              return reservedVerticalEndpointDirections.has(owner);
-            })
-          ) ? VERTICAL_GAP : 0
-        }
-      );
-
-      record.size = { width: placement.width, height: placement.height };
-      layout.shapes.set(record.element, placement);
-      placedArtifacts.push({
-        element: record.element,
-        rect: placement,
-        annotationClearance
-      });
-    }
-
-    for (const record of artifactRecords) {
-      const artifactBounds = layout.shapes.get(record.element);
-      const references = owners.get(record.element) || [];
-
-      for (const {
-        association,
-        owner,
-        ownerConnectionIndex,
-        ownerConnectionCount
-      } of references) {
-        layout.edges.set(association, routeArtifactAssociation(
-          association,
-          owner,
-          graphShapes.get(owner),
-          record.element,
-          artifactBounds,
-          ownerConnectionIndex,
-          ownerConnectionCount,
-          record.associationObstacles,
-          graphRoutes
-        ));
-      }
-    }
-
-  }
-
-  routeSequenceFlows(sequenceFlows, layout, policy) {
-    const routedConnections = [];
-    const shapes = [ ...layout.shapes.entries(), ...getExpandedChildShapes(layout) ]
-      .filter(([ element ]) => {
-        return !is(element, 'bpmn:Lane') &&
-          !is(element, 'bpmn:Participant') &&
-          !isArtifact(element);
-      })
-      .map(([ element, rect ]) => ({ element, rect }));
-
-    const ordered = [ ...sequenceFlows ].sort((a, b) => {
-      return edgePriority(a, policy) - edgePriority(b, policy) ||
-        policy.edgeOrder.get(a) - policy.edgeOrder.get(b);
-    });
-
-    for (const flow of ordered) {
-      const source = layout.shapes.get(flow.sourceRef);
-      const target = layout.shapes.get(flow.targetRef);
-
-      if (!source || !target) {
-        continue;
-      }
-
-      const points = routeConnection(flow, source, target, shapes, routedConnections, policy);
-      layout.edges.set(flow, points);
-      routedConnections.push({ flow, points });
-    }
   }
 
   emitDiagram(definitions, root, layout) {
@@ -920,54 +416,6 @@ export class Layouter {
       ));
     }
   }
-}
-
-export function collectArtifactObstacleRoutes(
-    layout,
-    annotatedMessageEndpoints,
-    graphShapes) {
-  const routes = [
-    ...layout.edges.entries(),
-    ...getExpandedChildEdges(layout)
-  ]
-    .filter(([ element ]) => {
-      return is(element, 'bpmn:SequenceFlow') || is(element, 'bpmn:MessageFlow');
-    })
-    .map(([ element, points ]) => ({ element, points }));
-  const reservedRoutes = [ ...annotatedMessageEndpoints ].flatMap(([
-    element,
-    directions
-  ]) => {
-    const rect = graphShapes.get(element);
-
-    if (!rect) {
-      return [];
-    }
-
-    const centerX = rect.x + rect.width / 2;
-
-    return [
-      directions.has('incoming') && {
-        element,
-        points: [
-          { x: centerX, y: rect.y },
-          { x: centerX, y: rect.y - MAX_ARTIFACT_SEARCH_OFFSET }
-        ]
-      },
-      directions.has('outgoing') && {
-        element,
-        points: [
-          { x: centerX, y: rect.y + rect.height },
-          {
-            x: centerX,
-            y: rect.y + rect.height + MAX_ARTIFACT_SEARCH_OFFSET
-          }
-        ]
-      }
-    ].filter(Boolean);
-  });
-
-  return [ ...routes, ...reservedRoutes ];
 }
 
 function compactParticipantRows(participants, participantShapes, participantLayouts) {
