+++ lib/layout/LayoutUtil.js
@@ -1,6 +1,16 @@
 import { OUTER_MARGIN, SEGMENT_INTERSECTION_EPSILON } from './Constants.js';
 import { isExteriorArtifact } from './BpmnUtil.js';
 
+/**
+ * @typedef {import('./Types.js').LayoutState} LayoutState
+ */
+
+/**
+ * Create the empty state for one independently laid-out BPMN scope.
+ *
+ * @param {Object} scope
+ * @returns {LayoutState}
+ */
 export function createLayout(scope) {
   return {
     scope,
